Added support for the GOG DVD version executable + fixes.
Made the patch compatible  with the "wc4dvd.exe" in the GOG  release of the game.
Fixed a problem with the offset size in palette updating functions.
Added music class destruction on exit to the Win95 Update version, which seemed to be missing.
Renamed some variables.
Fixed a potential problem with inflight movie playback.